Do not implement batches using transactions
Transactions have a size limit and will only accept a finite number of writes before returning ErrTxnTooBig.  Batching should not be subjected to this limit.

This PR fixes this by using a WriteBatch object to implement batching.  The WriteBatch object packs writes into transactions, committing the transactions as they become full.

This addresses Issue #21 by providing a batching mechanism that is not constrained to the limits of a single transaction.